Simplified if..else if in PluginFactory for Java plugins part, moved to template method pattern
