+++ logstash-core/lib/logstash/pipeline.rb
@@ -46,7 +46,7 @@ module LogStash; class BasePipeline < AbstractPipeline
 
     @plugin_factory = LogStash::Plugins::PluginFactory.new(
       # use NullMetric if called in the BasePipeline context otherwise use the @metric value
-      lir, LogStash::Plugins::PluginMetricFactory.new(pipeline_id, metric),
+      lir, LogStash::Plugins::PluginMetricsFactory.new(pipeline_id, metric),
       LogStash::Plugins::ExecutionContextFactory.new(@agent, self, dlq_writer),
       FilterDelegator
     )
