+++ logstash-core/src/main/java/org/logstash/execution/JavaBasePipelineExt.java
@@ -36,7 +36,9 @@ import org.logstash.common.IncompleteSourceWithMetadataException;
 import org.logstash.config.ir.CompiledPipeline;
 import org.logstash.execution.queue.QueueWriter;
 import org.logstash.ext.JRubyWrappedWriteClientExt;
-import org.logstash.plugins.PluginFactoryExt;
+import org.logstash.plugins.factory.ExecutionContextFactoryExt;
+import org.logstash.plugins.factory.PluginMetricsFactoryExt;
+import org.logstash.plugins.factory.PluginFactoryExt;
 
 import java.security.NoSuchAlgorithmException;
 import java.util.Collection;
@@ -67,12 +69,12 @@ public final class JavaBasePipelineExt extends AbstractPipelineExt {
         initialize(context, args[0], args[1], args[2]);
         lirExecution = new CompiledPipeline(
             lir,
-            new PluginFactoryExt.Plugins(context.runtime, RubyUtil.PLUGIN_FACTORY_CLASS).init(
+            new PluginFactoryExt(context.runtime, RubyUtil.PLUGIN_FACTORY_CLASS).init(
                 lir,
-                new PluginFactoryExt.Metrics(
-                    context.runtime, RubyUtil.PLUGIN_METRIC_FACTORY_CLASS
+                new PluginMetricsFactoryExt(
+                    context.runtime, RubyUtil.PLUGIN_METRICS_FACTORY_CLASS
                 ).initialize(context, pipelineId(), metric()),
-                new PluginFactoryExt.ExecutionContext(
+                new ExecutionContextFactoryExt(
                     context.runtime, RubyUtil.EXECUTION_CONTEXT_FACTORY_CLASS
                 ).initialize(context, args[3], this, dlqWriter(context)),
                 RubyUtil.FILTER_DELEGATOR_CLASS
