+++ logstash-core/src/main/java/org/logstash/plugins/codecs/Line.java
@@ -77,16 +77,26 @@ public class Line implements Codec {
     /**
      * Required constructor.
      *
+     * @param id            plugin id
+     * @param configuration Logstash Configuration
+     * @param context       Logstash Context
+     */
+    public Line(final String id, final Configuration configuration, final Context context) {
+        this(context, configuration.get(DELIMITER_CONFIG), configuration.get(CHARSET_CONFIG), configuration.get(FORMAT_CONFIG),
+                (id != null && !id.isEmpty()) ? id : UUID.randomUUID().toString());
+    }
+
+    /*
      * @param configuration Logstash Configuration
      * @param context       Logstash Context
      */
     public Line(final Configuration configuration, final Context context) {
-        this(context, configuration.get(DELIMITER_CONFIG), configuration.get(CHARSET_CONFIG), configuration.get(FORMAT_CONFIG));
+        this(null, configuration, context);
     }
 
-    private Line(Context context, String delimiter, String charsetName, String format) {
+    private Line(Context context, String delimiter, String charsetName, String format, String id) {
         this.context = context;
-        this.id = UUID.randomUUID().toString();
+        this.id = id;
         this.delimiter = delimiter;
         this.charset = Charset.forName(charsetName);
         this.format = format;
@@ -165,6 +175,6 @@ public class Line implements Codec {
 
     @Override
     public Codec cloneCodec() {
-        return new Line(context, delimiter, charset.name(), format);
+        return new Line(context, delimiter, charset.name(), format, UUID.randomUUID().toString());
     }
 }
