+++ logstash-core/src/main/java/org/logstash/plugins/factory/InputPluginCreator.java
@@ -1,0 +1,27 @@
+package org.logstash.plugins.factory;
+
+import co.elastic.logstash.api.Context;
+import co.elastic.logstash.api.Input;
+import org.jruby.runtime.builtin.IRubyObject;
+import org.logstash.config.ir.compiler.JavaInputDelegatorExt;
+import org.logstash.execution.JavaBasePipelineExt;
+import org.logstash.instrument.metrics.AbstractNamespacedMetricExt;
+import org.logstash.plugins.PluginLookup;
+
+import java.util.Map;
+
+class InputPluginCreator extends AbstractPluginCreator<Input> {
+
+    InputPluginCreator(PluginFactoryExt pluginsFactory) {
+        this.pluginsFactory = pluginsFactory;
+    }
+
+    @Override
+    public IRubyObject createDelegator(String name, Map<String, Object> pluginArgs, String id,
+                                       AbstractNamespacedMetricExt typeScopedMetric,
+                                       PluginLookup.PluginClass pluginClass, Context pluginContext) {
+        Input input = instantiateAndValidate(pluginArgs, id, pluginContext, pluginClass);
+        return JavaInputDelegatorExt.create((JavaBasePipelineExt) pluginsFactory.getExecutionContextFactory().getPipeline(),
+                typeScopedMetric, input, pluginArgs);
+    }
+}
